refactor: move Response namespace to ExplorerService
