+++ src/types/index.d.ts
@@ -610,31 +610,3 @@ declare namespace CustomRouter {
     comp: React.FunctionComponent<any>
   }
 }
-
-declare namespace Response {
-  export interface Response<T> {
-    data: T
-    meta?: Meta
-    error?: Error[]
-  }
-
-  export interface Error {
-    id: string
-    code: number
-    status: number
-    title: string
-    detail: string
-    href: string
-  }
-
-  export interface Meta {
-    total: number
-    pageSize: number
-  }
-
-  export interface Wrapper<T> {
-    id: number
-    type: string
-    attributes: T
-  }
-}
