+++ src/utils/hook.ts
@@ -25,6 +25,8 @@ import {
 } from './cache'
 import { parseDate } from './date'
 import { omit } from './object'
+// TODO: This file depends on higher-level abstractions, so it should not be in the utils folder. It should be moved to `src/hooks/index.ts`.
+import { Response } from '../services/ExplorerService'
 
 /**
  * Returns the value of the argument from the previous render
